fix: moved errors to its own folder
